requirements-test.txt
# requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pandas>=1.3.0
numpy>=1.20.0
PyYAML>=6.0
tabulate>=0.9.0